misc fixes

+++ repair.py
@@ -29,6 +29,8 @@ parser.add_argument('--no-confirm', action='store_true', help='Execute without c
 parser.add_argument('--repair-interval', type=str, default=repair['repairInterval'], help='Optional interval in smart format (e.g. 1h2m3s) to wait between repairing each media file.')
 parser.add_argument('--run-interval', type=str, default=repair['runInterval'], help='Optional interval in smart format (e.g. 1w2d3h4m5s) to run the repair process.')
 parser.add_argument('--mode', type=str, choices=['symlink', 'file'], default='symlink', help='Choose repair mode: `symlink` or `file`. `symlink` to repair broken symlinks and `file` to repair missing files.')
+parser.add_argument('--season-packs', action='store_true', help='Upgrade to season-packs when a non-season-pack is found. Only applicable in symlink mode.')
+parser.add_argument('--soft-repair', action='store_true', help='Only search for missing files, do not delete or re-grab. This is always enabled in file mode.')
 parser.add_argument('--include-unmonitored', action='store_true', help='Include unmonitored media in the repair process')
 args = parser.parse_args()
 
@@ -102,7 +104,7 @@ def main():
                     if args.dry_run or args.no_confirm or input("Do you want to delete and re-grab? (y/n): ").lower() == 'y':
                         if not args.dry_run:
                             discordUpdate(f"[{args.mode}] Repairing {media.title}: {childId}")
-                            if args.mode == 'symlink':
+                            if args.mode == 'symlink' and not args.soft_repair:
                                 print("Deleting files:")
                                 [print(item.path) for item in childItems]
                                 results = arr.deleteFiles(childItems)
@@ -127,9 +129,17 @@ def main():
                     if childId in media.fullyAvailableChildrenIds and len(parentFolders) > 1:
                         print("Title:", media.title)
                         print("Movie ID/Season Number:", childId)
-                        print("Inconsistent folders:")
+                        print("Non-season-pack folders:")
                         [print(parentFolder) for parentFolder in parentFolders]
                         print()
+                        if args.season_packs:
+                            print("Searching for season-pack")
+                            results = arr.automaticSearch(media, childId)
+                            print(results)
+
+                            if repairIntervalSeconds > 0:
+                                time.sleep(repairIntervalSeconds)
+
         except Exception:
             e = traceback.format_exc()
